Refactor: Remove presentation logic from coordinator
